+++ src/main/java/vn/hoidanit/jobhunter/controller/RoleController.java
@@ -24,7 +24,6 @@ import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestParam;
 
 @RestController
 @RequestMapping("/api/v1")
@@ -72,7 +71,11 @@ public class RoleController {
 
     @DeleteMapping("/roles/{id}")
     @ApiMessage("delete role")
-    public ResponseEntity<Void> deleteRole(@PathVariable long id) {
+    public ResponseEntity<Void> deleteRole(@PathVariable long id) throws IdInvalidException {
+
+        if (!this.roleService.existById(id)) {
+            throw new IdInvalidException("ID không tồn tại");
+        }
 
         this.roleService.deleteById(id);
